fix some bugs and embed cuda-runner(not finished yet)

+++ mlir-toy/include/toy/Lexer.h
@@ -85,7 +85,7 @@ public:
   }
 
   /// Return the current number (prereq: getCurToken() == tok_number)
-  double getValue() {
+  float getValue() {
     assert(curTok == tok_number);
     return numVal;
   }
@@ -193,7 +193,7 @@ private:
   std::string identifierStr;
 
   /// If the current Token is a number, this contains the value.
-  double numVal = 0;
+  float numVal = 0;
 
   /// The last value returned by getNextChar(). We need to keep it around as we
   /// always need to read ahead one character to decide when to end a token and
